Fix call forwarding and null handling in transforming collections
The iterator and spliterator used by 'AbstractTransformingCollection' did not immediately forward all transformation calls to the collection. This caused problems because the used (spl)iterator assumed a certain way of handling null, which the 'AbstractTransformingCollection' explicitly avoids. This brought forth that the 'AbstractTransformingMap.EntrySetView' must handle nulls itself and in a slightly different way.